// types/Piece.ts
enum Piece {
  HumanPawnA = "humanPawnA",
  HumanPawnB = "humanPawnB",
  HumanPawnC = "humanPawnC",
  HumanPawnD = "humanPawnD",
  HumanPawnE = "humanPawnE",
  HumanPawnF = "humanPawnF",
  HumanPawnG = "humanPawnG",
  HumanPawnH = "humanPawnH",
  HumanRookA = "humanRookA",
  HumanRookB = "humanRookB",
  HumanKnightA = "humanKnightA",
  HumanKnightB = "humanKnightB",
  HumanBishopA = "humanBishopA",
  HumanBishopB = "humanBishopB",
  HumanQueen = "humanQueen",
  HumanKing = "humanKing",
  AiPawnA = "aiPawnA",
  AiPawnB = "aiPawnB",
  AiPawnC = "aiPawnC",
  AiPawnD = "aiPawnD",
  AiPawnE = "aiPawnE",
  AiPawnF = "aiPawnF",
  AiPawnG = "aiPawnG",
  AiPawnH = "aiPawnH",
  AiRookA = "aiRookA",
  AiRookB = "aiRookB",
  AiKnightA = "aiKnightA",
  AiKnightB = "aiKnightB",
  AiBishopA = "aiBishopA",
  AiBishopB = "aiBishopB",
  AiQueen = "aiQueen",
  AiKing = "aiKing",
  None = "none",
}

export default Piece;